pyproject.toml
[tool.poetry]
name = "py-encryptor"
version = "0.1.0"
description = ""
authors = ["Krystian Postek <krystian@postek.eu>"]
#readme = "README.md"
packages = [{include = "py_encryptor"}]

[tool.poetry.dependencies]
python = "^3.11"
pycryptodomex = "^3.17"
pysimplegui = "^4.60.4"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pyencryptor = 'py_encryptor.cli:cli'
pyencryptor-gui = 'py_encryptor.gui:main'
